+++ src/decoder.h
@@ -29,6 +29,7 @@
 #define decoder_h
 
 #include <stddef.h>
+#include <stdint.h>
 #include <stdio.h>
 
 #include "wapiti.h"
@@ -36,9 +37,9 @@
 #include "sequence.h"
 
 void tag_viterbi(mdl_t *mdl, const seq_t *seq,
-	         size_t out[], double *sc, double psc[]);
-void tag_nbviterbi(mdl_t *mdl, const seq_t *seq, size_t N,
-	           size_t out[][N], double sc[], double psc[][N]);
+                 uint32_t out[], double *sc, double psc[]);
+void tag_nbviterbi(mdl_t *mdl, const seq_t *seq, uint32_t N,
+                   uint32_t out[][N], double sc[], double psc[][N]);
 
 void tag_label(mdl_t *mdl, FILE *fin, FILE *fout);
 void tag_eval(mdl_t *mdl, double *te, double *se);
