+++ src/gradient.h
@@ -43,26 +43,26 @@
 typedef struct grd_s grd_t;
 struct grd_s {
 	mdl_t    *mdl;
-	int       len;     // =T        max length of sequence
+	uint32_t len;     // =T        max length of sequence
 	double   *g;       // [F]       vector where to put gradient updates
 	double    lloss;   //           loss value for the sequence
 	double   *psi;     // [T][Y][Y] the transitions scores
 	double   *psiuni;  // [T][Y]    | Same as psi in sparse format
-	uint64_t *psiyp;   // [T][Y][Y] |
-	uint64_t *psiidx;  // [T][Y]    |
-	uint64_t *psioff;  // [T]
+	uint32_t *psiyp;   // [T][Y][Y] |
+	uint32_t *psiidx;  // [T][Y]    |
+	uint32_t *psioff;  // [T]
 	double   *alpha;   // [T][Y]    forward scores
 	double   *beta;    // [T][Y]    backward scores
 	double   *scale;   // [T]       scaling factors of forward scores
 	double   *unorm;   // [T]       normalization factors for unigrams
 	double   *bnorm;   // [T]       normalization factors for bigrams
-	int       first;   //           first position where gradient is needed
-	int       last;    //           last position where gradient is needed
+	uint32_t  first;   //           first position where gradient is needed
+	uint32_t  last;    //           last position where gradient is needed
 };
 
 grd_t *grd_new(mdl_t *mdl, double *g);
 void grd_free(grd_t *grd);
-void grd_check(grd_t *grd, int len);
+void grd_check(grd_t *grd, uint32_t len);
 
 void grd_fldopsi(grd_t *grd, const seq_t *seq);
 void grd_flfwdbwd(grd_t *grd, const seq_t *seq);
