+++ src/sequence.h
@@ -73,8 +73,8 @@
  */
 typedef struct raw_s raw_t;
 struct raw_s {
-	int   len;      //   T     Sequence length
-	char *lines[];  //  [T]    Raw lines directly from file
+	uint32_t  len;      //   T     Sequence length
+	char     *lines[];  //  [T]    Raw lines directly from file
 };
 
 /* tok_t:
@@ -92,10 +92,10 @@ struct raw_s {
  */
 typedef struct tok_s tok_t;
 struct tok_s {
-	int    len;     //   T     Sequence length
-	char **lbl;     //  [T]    List of labels strings
-	int   *cnts;    //  [T]    Length of tokens lists
-	char **toks[];  //  [T][]  Tokens lists
+	uint32_t   len;     //   T     Sequence length
+	char     **lbl;     //  [T]    List of labels strings
+	uint32_t  *cnts;    //  [T]    Length of tokens lists
+	char     **toks[];  //  [T][]  Tokens lists
 };
 
 /* seq_t:
@@ -122,10 +122,10 @@ struct tok_s {
 typedef struct pos_s pos_t;
 typedef struct seq_s seq_t;
 struct seq_s {
-	int       len;
+	uint32_t  len;
 	uint64_t *raw;
 	struct pos_s {
-		uint64_t  lbl;
+		uint32_t  lbl;
 		uint32_t  ucnt,  bcnt;
 		uint64_t *uobs, *bobs;
 	} pos[];
@@ -140,10 +140,10 @@ struct seq_s {
  */
 typedef struct dat_s dat_t;
 struct dat_s {
-	bool     lbl;   //         True iff sequences are labelled
-	int      mlen;  //         Length of the longest sequence in the set
-	size_t   nseq;  //   S     Number of sequences in the set
-	seq_t  **seq;   //  [S]    List of sequences
+	bool       lbl;   //         True iff sequences are labelled
+	uint32_t   mlen;  //         Length of the longest sequence in the set
+	uint32_t   nseq;  //   S     Number of sequences in the set
+	seq_t    **seq;   //  [S]    List of sequences
 };
 
 #endif
