+++ src/thread.h
@@ -28,6 +28,7 @@
 #ifndef thread_h
 #define thread_h
 
+#include <stdint.h>
 #include <pthread.h>
 
 #include "model.h"
@@ -36,7 +37,7 @@ typedef struct job_s job_t;
 
 typedef void (func_t)(job_t *job, int id, int cnt, void *ud);
 
-bool mth_getjob(job_t *job, size_t *cnt, size_t *pos);
-void mth_spawn(func_t *f, int W, void *ud[W], size_t size, size_t batch);
+bool mth_getjob(job_t *job, uint32_t *cnt, uint32_t *pos);
+void mth_spawn(func_t *f, int W, void *ud[W], uint32_t size, uint32_t batch);
 
 #endif
